replaced and renamed some of the components for consistency
